+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- changed `persisted_flag_callback_control` default from `false` to `true`.
 - adjusted to save errors as soon as an invalid is found.
 - drop support rails 6.1.x.